+++ src/operations/getDocs.test.ts
@@ -6,7 +6,7 @@ import {
 	userRefCreator,
 	User,
 } from '../utilForTests'
-import { setDoc } from '../operations'
+import { setDoc } from './setDoc'
 import {
 	IsSame,
 	IsTrue,
@@ -148,10 +148,10 @@ describe('test getDocs', () => {
 		const d2 = generateRandomData()
 		const d3 = generateRandomData()
 		const d4 = generateRandomData()
-		const p1 = setDoc(userRef.doc('1'), d1)
-		const p2 = setDoc(userRef.doc('2'), d2)
-		const p3 = setDoc(userRef.doc('3'), d3)
-		const p4 = setDoc(userRef.doc('4'), d4)
+		const p1 = setDoc(userRef.doc('getDocsCursorTest1'), d1)
+		const p2 = setDoc(userRef.doc('getDocsCursorTest2'), d2)
+		const p3 = setDoc(userRef.doc('getDocsCursorTest3'), d3)
+		const p4 = setDoc(userRef.doc('getDocsCursorTest4'), d4)
 
 		await Promise.all([p1, p2, p3, p4])